requirements.txt
Flask
libvirt-python
prometheus_client
six
gevent >= 1.1.0
wsgigzip >= 0.1.4
python-dateutil
